Move layers list to components directory
