--- napari/components/_layers_list/__init__.py
@@ -1,1 +1,0 @@
-from .layers_list import LayersList
